+++ src/server.ts
@@ -1,5 +1,7 @@
 import * as vscode from 'vscode';
 
+import ip from 'ip';
+
 import Koa, { Request } from 'koa';
 import Router from 'koa-router';
 
@@ -10,6 +12,8 @@ import mount from 'koa-mount';
 
 import { getPort, setBasePort, setHighestPort } from 'portfinder';
 
+// Reference to create ssl certificates for https: https://stackoverflow.com/questions/10175812/how-to-generate-a-self-signed-ssl-certificate-using-openssl
+import https from 'https';
 import http from 'http';
 import path from 'path';
 import fs from 'fs';
@@ -21,8 +25,9 @@ export default class KoaApp {
   private _instance: Koa;
   private _router: Router;
   private _serverApp: http.Server<any>;
-	private _serverPortOptions: number[] = [ 8090, 9000 ];
-  private _whitelist = [ 'https://test2.voly.co.uk', 'http://localhost', 'voly.docker' ];
+  private _securedServerApp: https.Server<any>;
+	private _serverPortOptions: number[] = [ 8090, 9999 ];
+  private _whitelist = [ 'https://test2.voly.co.uk', 'http://localhost', 'http://voly.docker' ];
   private _corsOptions: cors.Options = {
     origin: (request: Request) => {
       if (this._whitelist.indexOf(request.headers.origin || '') > -1 || !request.headers.origin) {
@@ -48,6 +53,11 @@ export default class KoaApp {
     setHighestPort(this._serverPortOptions[1]);
 
     this._serverApp = http.createServer(this._instance.callback());
+    this._securedServerApp = https.createServer({
+      key: fs.readFileSync(path.join(__dirname, '..', 'src', 'ssl-certs', 'key.pem'), 'utf8').toString(),
+      cert: fs.readFileSync(path.join(__dirname, '..', 'src', 'ssl-certs', 'cert.pem'), 'utf8').toString(),
+    }, this._instance.callback());
+
     getPort((err, port) => {
       if (err) {
         vscode.window.showErrorMessage(`[volyfequickdev] ${err.message}`);
@@ -57,22 +67,30 @@ export default class KoaApp {
     });
   }
 
+  // non-secure server
   public checkListeningState() {
-    return this._serverApp.listening;
+    return this._serverApp?.listening;
   }
 
   public startServer(port: number) {
+    // Stop the operation if server is already listening on a port
+    if (this.checkListeningState()) {
+      console.warn('Local server is currently already listening');
+      return;
+    }
+    // Check if the secured server is listening and then close it
+    this.closeSecuredServer();
     this.selectedPort = port || this.selectedPort;
     this._serverApp.listen(this.selectedPort, async () => {
       await vscode.commands.executeCommand('setContext', 'volyfequickdev-settings.serverIsOn', true);
-      vscode.window.showInformationMessage(`[volyfequickdev] Server is running on port ${port}`);
+      vscode.window.showInformationMessage(`[volyfequickdev] Server is running on http://localhost:${port}`);
     });
   }
 
-  public closeServer() {
+  public closeServer(): boolean {
     if (!this.checkListeningState()) {
       console.warn('Server is no longer listening');
-      return;
+      return false;
     }
     this._serverApp.close(async (err) => {
       if (err) {
@@ -81,6 +99,42 @@ export default class KoaApp {
       }
       await vscode.commands.executeCommand('setContext', 'volyfequickdev-settings.serverIsOn', false);
     });
+    return true;
+  }
+
+  // secure server
+  public checkSecuredListeningState() {
+    return this._securedServerApp?.listening;
+  }
+
+  public startSecuredServer(port: number) {
+    // Stop the operation if server is already listening on a port
+    if (this.checkSecuredListeningState()) {
+      console.warn('Secured server is currently already listening');
+      return;
+    }
+    // Check if the localhost server is listening and then close it
+    this.closeServer();
+    this.selectedPort = port || this.selectedPort;
+    this._securedServerApp.listen(this.selectedPort, async () => {
+      await vscode.commands.executeCommand('setContext', 'volyfequickdev-settings.serverIsOn', true);
+      vscode.window.showInformationMessage(`[volyfequickdev] Server is running on https://${ip.address()}:${port}`);
+    });
+  }
+
+  public closeSecuredServer(): boolean {
+    if (!this.checkSecuredListeningState()) {
+      console.warn('Secured server is no longer listening');
+      return false;
+    }
+    this._securedServerApp.close(async (err) => {
+      if (err) {
+        vscode.window.showErrorMessage(`[volyfequickdev] ${err.message}`);
+        process.exit();
+      }
+      await vscode.commands.executeCommand('setContext', 'volyfequickdev-settings.serverIsOn', false);
+    });
+    return true;
   }
 
   public serveStatic() {